Moved code out of src
