+++ Application/GameWindowHandler.h
