+++ Engine/Objects/Player.h
@@ -13,7 +13,7 @@
 #include "Engine/Objects/ItemEnums.h"
 #include "Engine/Objects/PlayerEnums.h"
 #include "Engine/Spells/Spells.h"
-#include "src/tools/Workaround.h"
+#include "Utility/Workaround.h"
 
 
 struct LloydBeacon {
