+++ GUI/CMakeLists.txt
@@ -76,4 +76,4 @@ endif()
 add_library(gui STATIC ${GUI_SOURCES} ${GUI_HEADERS})
 target_check_style(gui)
 
-target_link_libraries(gui arcomage)
+target_link_libraries(gui arcomage utility)
