+++ GUI/UI/UIGame.cpp
@@ -4,7 +4,7 @@
 
 #include "GUI/UI/UIGame.h"
 
-#include "src/Application/Game.h"
+#include "Application/Game.h"
 
 #include "Engine/AssetsManager.h"
 #include "Engine/Engine.h"
