+++ GUI/UI/UIGuilds.cpp
@@ -21,7 +21,7 @@
 
 #include "Media/Audio/AudioPlayer.h"
 
-#include "src/tools/Workaround.h"
+#include "Utility/Workaround.h"
 
 using EngineIoc = Engine_::IocContainer;
