--- src/CMakeLists.txt
@@ -1,30 +1,0 @@
-cmake_minimum_required(VERSION 3.13 FATAL_ERROR)
-
-set(GAME_SOURCES Application/Game.cpp
-                 Application/GameBootstrapper.cpp
-                 Application/GameConfig.cpp
-                 Application/GameFactory.cpp
-                 Application/GameMenu.cpp
-                 Application/GameOver.cpp
-                 Application/GameWindowHandler.cpp
-                 Application/IocContainer.cpp
-                 tools/DataPath.cpp
-                 tools/MemBuffer.cpp)
-
-set(GAME_HEADERS Application/Game.h
-                 Application/GameConfig.h
-                 Application/GameFactory.h
-                 Application/GameMenu.h
-                 Application/GameOver.h
-                 Application/IocContainer.h
-                 Application/GameWindowHandler.h
-                 tools/Flags.h
-                 tools/DataPath.h
-                 tools/MemBuffer.h
-                 tools/FreeDeleter.h
-                 tools/Workaround.h)
-
-add_library(game STATIC ${GAME_SOURCES} ${GAME_HEADERS})
-target_check_style(game)
-
-target_link_libraries(game engine platform graphics)
